Move map-related TS files to map dir
